+++ web-ui-nuxt/api/get_sample.py
@@ -1,30 +1,28 @@
-from supabase import create_client, Client
-from dotenv import load_dotenv
-import os
 import random
-from api.PhonemeConverter import get_phonetic_transcription
-
-# Загрузка переменных окружения
-load_dotenv()
-sb_url: str = os.environ.get("SUPABASE_URL")
-sb_key: str = os.environ.get("SUPABASE_KEY")
-
-# Создание клиента Supabase
-supabase: Client = create_client(supabase_url=sb_url, supabase_key=sb_key)
+from api.phoneme_converter import get_phonetic_transcription
+from api.supabase_queries import get_sentences_with_length
 
 
 def get_sample_response(sentence_length_group: str) -> dict:
     """
     Возвращает JSON ответ со случайным предложением и его транскрипцией.
     """
-    # Получаем данные из таблицы Text_data
-    data = supabase.table("Text_data").select("*").execute()
+    # Получаем данные из таблицы с предложениями напрямую из бд (Supabase)
+    data = get_sentences_with_length(
+        sentence_length_group
+    )
+    
+    print("[DEBUG]", type(data), type(data.data))
 
     # Преобразуем данные в список словарей
     records = data.data  # Содержимое данных будет доступно через data.data
 
     # Фильтруем записи по категории sentence_length_group
-    filtered_records = [record for record in records if record['sentence_length_group'] == sentence_length_group]
+    filtered_records = [
+        record
+        for record in records
+        if record["sentence_length_group"] == sentence_length_group
+    ]
 
     if not filtered_records:
         return {"error": "No data available for the specified category"}
@@ -41,7 +39,7 @@ def get_sample_response(sentence_length_group: str) -> dict:
     response_data = {
         "en_sentence": en_sentence,
         "sentence_length_group": sentence_length_group,
-        "phonetic_transcription": phonetic_transcription
+        "phonetic_transcription": phonetic_transcription,
     }
 
     return response_data
