+++ web-ui-nuxt/api/index.py
@@ -1,8 +1,7 @@
 from fastapi import FastAPI, HTTPException
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
-
-from api.GetSample import get_sample_response
+from api.get_sample import get_sample_response
 
 app = FastAPI()
 
@@ -11,12 +10,14 @@ class SentenceRequest(BaseModel):
     sentence_length_group: str
 
 
-@app.post("/api/getSample")
+@app.post("/api/get_sample")
 async def get_sample(request: SentenceRequest):
     response_data = get_sample_response(request.sentence_length_group)
 
     if not response_data:
-        raise HTTPException(status_code=404, detail="No data available for the specified category")
+        raise HTTPException(
+            status_code=404, detail="No data available for the specified category"
+        )
 
     return JSONResponse(content=response_data)
 
@@ -24,9 +25,11 @@ async def get_sample(request: SentenceRequest):
 @app.get("/transcribe")
 async def transcribe():
     """
-    Возвращает фонемную транскрипцию аудиофайла test.wav.
+    Возвращает фонемную транскрипцию аудио файла test.wav.
     """
-    file_path = "C:/Users/Asus/Desktop/ASR-model/test.wav"  # Укажите путь к вашему аудиофайлу
+    file_path = (
+        "C:/Users/Asus/Desktop/ASR-model/test.wav"  # Укажите путь к вашему аудио файлу
+    )
     transcription = transcribe_audio(file_path)
     return {"transcription": transcription}
